Moved logic code to the scripts. creating many global variables. Will try to minimize these
